requirements.txt
paho-mqtt==1.5.1
pycryptodome==3.9.9
